+++ packages/pluggableWidgets/color-picker-native/src/__tests__/ColorPicker.spec.tsx
@@ -1,4 +1,4 @@
-import { actionValue, EditableValueBuilder } from "@native-mobile-resources/util-widgets";
+import { actionValue, EditableValueBuilder } from "@widgets-resources/piw-utils";
 import { createElement } from "react";
 import { View } from "react-native";
 import Slider from "react-native-slider";
