+++ packages/pluggableWidgets/toggle-buttons-native/src/__tests__/ToggleButtons.spec.tsx
@@ -1,4 +1,4 @@
-import { actionValue, EditableValueBuilder } from "@native-mobile-resources/util-widgets";
+import { actionValue, EditableValueBuilder } from "@widgets-resources/piw-utils";
 import { createElement } from "react";
 import { Text } from "react-native";
 import { fireEvent, render } from "react-native-testing-library";
