+++ packages/tools/piw-utils/src/builders/EditableValueBuilder.ts
@@ -1,19 +1,8 @@
-import { ActionValue, DynamicValue, ValueStatus } from "mendix";
+import { ValueStatus } from "mendix";
 
-export function dynamicValue<T>(value?: T, loading?: boolean): DynamicValue<T> {
-    if (loading) {
-        return { status: ValueStatus.Loading, value };
-    }
-    return value ? { status: ValueStatus.Available, value } : { status: ValueStatus.Unavailable, value: undefined };
-}
+type Option<T> = T | undefined;
 
-export function actionValue(canExecute = true, isExecuting = false): ActionValue {
-    return { canExecute, isExecuting, execute: jest.fn() };
-}
-
-declare type Option<T> = T | undefined;
-
-declare interface EditableValue<T> {
+interface EditableValue<T> {
     status: ValueStatus;
     readOnly: boolean;
     value: Option<T>;
