+++ packages/tools/piw-utils/src/builders/index.ts
@@ -1,1 +1,3 @@
 export * from "./ListValueBuilder";
+export * from "./DynamicActionValueBuilder";
+export * from "./EditableValueBuilder";
