+++ packages/tools/util-widgets/src/index.ts
@@ -1,4 +1,4 @@
 export { flattenStyles, Style } from "./styles";
 export { only, exclude } from "./objectFilters";
 export { available, toNumber, unavailable } from "./properties";
-export * from "./test";
+export * from "../../piw-utils/src/builders/DynamicActionValueBuilder";
